+++ pypdf/annotations/__init__.py
@@ -1,0 +1,28 @@
+"""PDF annotations"""
+
+
+from ._base import NO_FLAGS
+from ._markup_annotations import (
+    Ellipse,
+    FreeText,
+    Highlight,
+    Line,
+    Link,
+    Polygon,
+    PolyLine,
+    Rectangle,
+    Text,
+)
+
+__all__ = [
+    "Line",
+    "Text",
+    "FreeText",
+    "PolyLine",
+    "Rectangle",
+    "Highlight",
+    "Ellipse",
+    "Polygon",
+    "Link",
+    "NO_FLAGS",
+]
